lib/stripe: Disable state updates when unmounted
If using the useStripeJs or useStripeConfiguration custom hooks or the
withStripe HOC that uses them, it's possible that the base component
could be unmounted before one of the async state updates completes (eg:
loading stripe.js).

Updating the state of an unmounted component is harmless but generates a
warning from React. This change adds a subscription variable to be sure
the component is mounted before setting the state.